Use `static_cast` when saving `mConnectorDirection` field
There was an implicit cast to `int` here, which would no longer be implicit if the `enum` was switched to `enum class`.

+++ appOPHD/MapObjects/Structure.cpp
@@ -495,7 +495,7 @@ NAS2D::Dictionary Structure::getDataDict() const
 		{"disabled_reason", static_cast<int>(mDisabledReason)},
 		{"idle_reason", static_cast<int>(mIdleReason)},
 		{"type", static_cast<std::size_t>(mStructureId)},
-		{"direction", mConnectorDirection},
+		{"direction", static_cast<int>(mConnectorDirection)},
 		{"integrity", mIntegrity},
 		{"pop0", mPopulationAvailable.workers},
 		{"pop1", mPopulationAvailable.scientists},
